Moving handleLearnLink to its own file
