+++ src/lib/remark-plugins/rewrite-tutorial-links/index.ts
@@ -27,9 +27,8 @@ import {
 	getIsDocsLink,
 	getIsLearnLink,
 	getTutorialMap,
-	handleCollectionLink,
 	handleDocsLink,
-	handleTutorialLink,
+	handleLearnLink,
 } from './utils'
 
 let TUTORIAL_MAP
@@ -45,51 +44,6 @@ export const rewriteTutorialLinksPlugin: Plugin = () => {
 
 function handleRewriteTutorialsLink(node: Link | Definition) {
 	node.url = rewriteTutorialsLink(node.url, TUTORIAL_MAP)
-}
-
-const handleLearnLink = (
-	urlObject: URL,
-	tutorialMap: Record<string, string>
-) => {
-	let newUrl
-
-	const { pathname } = urlObject
-	const pathnameParts = pathname.split('/')
-
-	let product
-	if (pathnameParts.length === 2) {
-		product = pathnameParts[1]
-	} else if (
-		pathnameParts[1] === 'tutorials' ||
-		pathnameParts[1] === 'collections'
-	) {
-		product = pathnameParts[2]
-	}
-
-	const isBetaProduct = getIsBetaProduct(product)
-	if (isBetaProduct) {
-		// Regexes for each path type
-		const collectionPathRegex = new RegExp('^/collections')
-		const tutorialPathRegex = new RegExp('^/tutorials')
-		const productHubPathRegex = new RegExp(`^/${product}/?$`)
-
-		// Derived path type booleans
-		const isCollectionPath = collectionPathRegex.test(pathname)
-		const isProductHubPath = productHubPathRegex.test(pathname)
-		const isTutorialPath = tutorialPathRegex.test(pathname)
-
-		if (isCollectionPath) {
-			newUrl = handleCollectionLink(urlObject)
-		} else if (isTutorialPath) {
-			newUrl = handleTutorialLink(urlObject, tutorialMap)
-		} else if (isProductHubPath) {
-			newUrl = `/${product}/tutorials`
-		}
-	} else {
-		newUrl = urlObject.toString()
-	}
-
-	return newUrl
 }
 
 export function rewriteTutorialsLink(
