+++ src/lib/remark-plugins/rewrite-tutorial-links/utils/index.ts
@@ -3,4 +3,5 @@ export { getIsLearnLink } from './get-is-learn-link'
 export { getTutorialMap } from './get-tutorial-map'
 export { handleCollectionLink } from './handle-collection-link'
 export { handleDocsLink } from './handle-docs-link'
+export { handleLearnLink } from './handle-learn-link'
 export { handleTutorialLink } from './handle-tutorial-link'
